refactor(deployments): moved some components out of root deployment directory
